got countdown working
