+++ app.js
@@ -1,24 +1,18 @@
+import compareNumbers from './compareNumbers.js'
+
 const button = document.getElementById('button');
 const guessesRemaining = document.getElementById('guessesRemaining');
-let userNumber = document.getElementById('userNumber');
-const correctNumber = 13;
+const winLose = document.getElementById('winLose');
+const lose = 'You ran out of guesses. You lose.';
 let count = 4;
-
-// function to compare correctNumber to userNumber
-function compareNumbers() {
-    if (userNumber === correctNumber) {
-        return 'Congratulations! You guessed the number!';
-    } else if (userNumber > correctNumber) {
-        return 'Your guess was too high--try again!';
-    } else if (userNumber < correctNumber) {
-        return 'Your guess was too low--try again!';
-    }
-};
 
 // function carried out on click. Needs to 1. Reduce counter; 2. Test the number; 3. Show result
 button.addEventListener('click', () => {
     console.log('Button works!');
-    count--;
     compareNumbers();
+    count--;
     guessesRemaining.textContent = count;
+    if (count === 0) {
+        winLose.textContent = lose;
+    }
 })
