+++ index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="main.css">
     <title>Guessing Game</title>
 </head>
 
@@ -24,7 +24,7 @@
         
         <input id="userNumber" type="number" min="1" max="20">
         <button id="button">Submit Guess</button><br><br>
-        Guesses remaining: <span id="guessesRemaining"></span>
+        Guesses remaining: <span id="guessesRemaining">4</span><br><br>
         <span id="numRel"></span>
         <span id="winLose"></span>
     </p>
